requirements.txt
anyio==4.4.0
APScheduler==3.6.3
asttokens==2.4.1
cachetools==4.2.2
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
django-environ==0.11.2
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.36
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
mypy==1.10.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.9.1
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
pre-commit==3.7.1
prompt_toolkit==3.0.47
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
ruff==0.4.8
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
sympy==1.12.1
tbb==2021.12.0
thop==0.1.1.post2209072238
tomli==2.0.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ultralytics==8.2.31
ultralytics-thop==0.2.8
urllib3==2.2.1
virtualenv==20.26.2
wcwidth==0.2.13
zipp==3.19.2
